refactor(spa): split register-modules.tsx into register-modules/ subdir
The 470-line register-modules.tsx grew over time into a god file mixing
orchestration, the Editor module definition, fs backend wiring, and inline
file-opener registrations. Subsequent P1 tasks (1.3 file-opener migration,
1.4 HMR dispose) need a stable home for editor-owned contributions, so split
the file now while behavior is unchanged:

- register-modules/editor-module.tsx — editorModuleDefinition (panes +
  declarative settings) and registerEditorFileOpeners() wrapping the three
  inline registerFileOpener calls. Task 1.3 will promote them into
  editorModuleDefinition.fileOpeners.
- register-modules/fs-backends.tsx — InApp/Daemon/Local backend registration,
  exported as registerBuiltinFsBackends(caps).
- register-modules/index.tsx — orchestrator that wires everything together,
  the four Wrapper components, the HMR dispose hook, and the dispatch /
  baseline tail. Re-exports dispatchSettingsContributions for callers.
- register-modules.tsx — transitional re-export shim so existing import paths
  keep working; lint disable retained for eslint-plugin-react-refresh which
  cannot reason through `export *`.

Also fix FileTreeView.test.tsx mock to declare ownerModuleId, required by the
RegisteredOpener type introduced in Task 1.0a.

No behavior change: SPA build, lint, and the full vitest suite are clean
relative to the pre-existing baseline failures (TabBar pinned-tab tooltip
and three sync hosts token-preservation cases) which are unrelated to P1.
